signficant speedup in highlighter by compiling the regex once
i thought my little lazy regex compiling caching thing worked but
apparently not and the editor was still recompiling the regular
expressions every frame... they are now compiled when the editor is
loaded instead. this is ideal since the editor will crash on a bad
config rather than encountering a highlight and crashing then.

and obviously its a lot faster too! its hard to gauge the speed increase
since im running this editor in a VM, but the fps went from 30 to 60
frames per second after doing this so that seems good to me!